Replace separator before files are passed to getNav
File paths was transformed with replaceDirsSeparator in the function
getNav but were not used in all following function calls.

Example: When filterFiles was called the files variable was passed.
navFiles = filterFiles(files, filterProperty, filterValue);
Result: navFiles back to backslash again.

Now the navFiles variable is transformed before it is passed to getNav.